src/functions/requirements.txt
azure-functions==1.18.0
azure-identity==1.15.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3